+++ quality_hub/src/App.js
@@ -3,12 +3,12 @@ import { Route } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import PrivateRoute from './utils/PrivateRoute';
-import LoginForm from './components/LoginForm';
-import SignUpForm from './components/SignUpForm';
+import LoginForm from './components/Forms/LoginForm';
+import SignUpForm from './components/Forms/SignUpForm';
 import Dashboard from './components/Dashboard';
 import Landing from './components/Landing';
-import InterviewerForm from './components/InterviewerForm';
-import StudentForm from './components/StudentForm';
+import InterviewerForm from './components/Forms/InterviewerForm';
+import StudentForm from './components/Forms/StudentForm';
 import UserType from './components/UserType/UserType';
 
 function App() {
