+++ quality_hub/src/components/Forms/InterviewerForm.jsx
@@ -4,14 +4,14 @@ import styled from 'styled-components';
 import { withFormik, Form, Field, FieldArray } from 'formik';
 import * as Yup from 'yup';
 
-import { StyledButton, buttonTheme, invertTheme } from './Landing';
+import { StyledButton, buttonTheme, invertTheme } from '../Landing';
 import {
   GreyBackgroundContainer,
   FormCard,
   FormContainer,
 } from './LoginForm';
 
-import { register } from '../state/actions/actionCreators';
+import { register } from '../../state/actions/actionCreators';
 
 const RegisterCard = styled(FormCard)`
   width: 30em;
