+++ quality_hub/src/components/Forms/LoginForm.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 import { withFormik, Form, Field } from 'formik';
-import { StyledButton, buttonTheme, invertTheme } from './Landing';
+import { StyledButton, buttonTheme, invertTheme } from '../Landing';
 import * as yup from 'yup';
 import axios from 'axios';
 
-import { login } from '../state/actions/actionCreators';
+import { login } from '../../state/actions/actionCreators';
 
 export const GreyBackgroundContainer = styled.div`
   height: 100vh;
